+++ src/test/kotlin/cl/ravenhill/jakt/constraints/ints/BeNegativeTest.kt
@@ -1,5 +1,6 @@
 package cl.ravenhill.jakt.constraints.ints
 
+import cl.ravenhill.jakt.assertions.constraints.`test BeNegative constraint`
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.booleans.shouldBeFalse
 import io.kotest.matchers.booleans.shouldBeTrue
@@ -10,22 +11,5 @@ import io.kotest.property.checkAll
 
 class BeNegativeTest : FreeSpec({
 
-    "A [BeNegative] constraint" - {
-        "should have a [validator] that" - {
-            "returns true if the value is negative" {
-                checkAll(Arb.negativeInt()) { value ->
-                    BeNegative.validator(value).shouldBeTrue()
-                }
-            }
-
-            "returns false if the value is non-negative" {
-                checkAll(Arb.positiveInt()) { value ->
-                    BeNegative.validator(value).shouldBeFalse()
-                }
-
-                // Checking zero as it's a boundary condition
-                BeNegative.validator(0).shouldBeFalse()
-            }
-        }
-    }
+    include(`test BeNegative constraint`(Arb.negativeInt(), Arb.positiveInt()) { BeNegative })
 })